Clear old messages on update. Remove caching.
The old messages need to always be cleared on an update.
This prevents fixed problems from being seen until a reload.

The caching of old errors is unnecessary. Users always want
the latest errors to be shown only. Removing the cache
reduces the complexity of the code to maintain.

With this, I also updated a parameter name from `doc` to
`updatedDocument` which makes the method far easier to read.